refresh and lunch

+++ main.sh
@@ -2,35 +2,53 @@
 . style.sh
 
 function main() {
-  echo $RED
   DIR=$(echo $3 | sed 's/%20/-/g' | cut -d' ' -f2)
-  echo "Cloning down $DIR."
-  git clone --single-branch --branch $4 "https://git.generalassemb.ly/$2/$1.git" --quiet $DIR
-  cd $DIR
-
-  echo "Checking dependencies for $DIR."
-  NPMEXISTS=$(find . -name 'package.json' ! -path '*node_modules*')
-  GEMEXISTS=$(find . -name 'Gemfile')
-
-  if [ -z "$NPMEXISTS" ]
+  REPOEXISTS=$(ls | grep $DIR)
+  if [ "$5" == "open" ]
   then
+    if [ -z "$REPOEXISTS" ]
+    then
+      # CLONE
+      echo "Cloning down $DIR."
+      git clone --single-branch --branch $4 "https://git.generalassemb.ly/$2/$1.git" --quiet $DIR
+      cd $DIR
+    else
+      # PULL FROM CURRENT BRANCH
+      cd $DIR
+      git pull origin $4
+    fi
+
+    echo "Checking dependencies for $DIR."
+    NPMEXISTS=$(find . -name 'package.json' ! -path '*node_modules*')
+    GEMEXISTS=$(find . -name 'Gemfile')
+
+    if [ -z "$NPMEXISTS" ]
+    then
     echo "No package.json detected.$RED Skipping installation.$RESET"
-  else
-    echo "Installing JS dependencies."
-    echo $NPMEXISTS |\
-      xargs -I '{}' sh -c "cd $(dirname '{}') && npm i --quiet && cd -"
+    else
+      echo "Installing JS dependencies."
+      echo $NPMEXISTS |\
+        xargs -I '{}' sh -c "cd $(dirname '{}') && npm i --quiet && cd -"
+    fi
+
+    if [ -z "$GEMEXISTS" ]
+    then
+      echo "No Gemfile detected.$RED Skipping installation.$RESET"
+    else
+      echo "Installing Ruby dependencies."
+      echo $GEMEXISTS |\
+        xargs -I '{}' sh -c "cd $(dirname '{}') && bundle install --quiet && cd -"
+    fi
+
+    cd ..
   fi
 
-  if [ -z "$GEMEXISTS" ]
+  if [ "$5" == "closed" ] && [ -n "$REPOEXISTS" ]
   then
-    echo "No Gemfile detected.$RED Skipping installation.$RESET"
-  else
-    echo "Installing Ruby dependencies."
-    echo $GEMEXISTS |\
-    xargs -I '{}' sh -c "cd $(dirname '{}') && bundle install --quiet && cd -"
+    echo "$DIR's homework is closed, removing directory."
+    rm -rf $DIR
   fi
 
-  cd ..
 }
 
 export -f main
@@ -45,13 +63,32 @@ fi
 printf $GREEN
 printf "${OPENER}"
 printf $RESET
+REFRESH=""
 
 echo "Welcome to Git Over Here,$BLUE $USER$RESET!"
 echo "You're currently pulling from $GREEN$COHORT$RESET."
-echo "What$GREEN repos$RESET would you like to grade? (e.g.$GREEN sequelize-pizza-express-routes rails-books-hw candies$RESET)"
-read INPUT
 
-REPOS=( $INPUT )
+while [[ "$1" =~ ^- && ! "$1" == "--" ]];
+do
+  case $1 in
+    -l | --lunch)
+      . .env.lunch
+      echo "Repos identified. Have a great lunch!"
+      ;;
+    -r | --refresh)
+      REFRESH="true"
+      ;;
+  esac
+  shift
+done
+
+if [ -z "$REPOS" ]
+then
+  echo "What$GREEN repos$RESET would you like to grade? (e.g.$GREEN sequelize-pizza-express-routes rails-books-hw candies$RESET)"
+  read INPUT
+  REPOS=( $INPUT )
+fi
+
 for REPO in "${REPOS[@]}"
 do
   SUCCESS=$(curl "https://git.generalassemb.ly/api/v3/repos/$COHORT/$REPO/pulls" -H "Authorization: token $TOKEN" |\
@@ -63,8 +100,11 @@ do
     echo "Repository not found. Please try again."
   else
     cd ..
-    rm -rf $REPO
-    mkdir $REPO
+    if [ -z "$REFRESH" ] || [ -z "$(ls | grep $REPO)" ]
+    then
+      rm -rf $REPO
+      mkdir $REPO
+    fi
     cd $REPO
 
     printf "$GREEN"
@@ -72,9 +112,9 @@ do
     printf "$RESET"
     OPENPRS="$(curl https://git.generalassemb.ly/api/v3/repos/$COHORT/$REPO/pulls\?state\=all -H "Authorization: token $TOKEN" | jq '.[] | .state' | grep "open" | wc -l | awk '$1=$1')"
     echo "You have $OPENPRS submissions, cloning into$BLUE $REPO$RESET."
-    curl "https://git.generalassemb.ly/api/v3/repos/$COHORT/$REPO/pulls" -H "Authorization: token $TOKEN" |\
-      jq '.[] | @uri "\(.user.login) \(.title) \(.head.ref)"' |\
-      xargs -L 3 -I {} sh -c "main $REPO {}"
+    curl "https://git.generalassemb.ly/api/v3/repos/$COHORT/$REPO/pulls?state=all" -H "Authorization: token $TOKEN" |\
+      jq '.[] | @uri "\(.user.login) \(.title) \(.head.ref) \(.state)"' |\
+      xargs -L 4 -I {} sh -c "main $REPO {}"
   fi
 
   printf "$RESET"
